pyproject.toml
[tool.poetry]
name = "langchain-project"
version = "0.1.0"
description = ""
authors = ["Chris Mayer"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
langchain = "^0.0.311"
openai = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
